Conflicts:
- gerbil_atlas_explorer.html, index.html: build outputs, both conflicted
  on nearly every line (this branch predates #48/#49/#50). Took main's
  side to clear the markers, then rebuilt both from the merged src/ and
  data/ with `python tools/build_app.py --lean` and verified with
  --check. Confirmed both the local-tz stamp feature and main's fissures
  fix (cbw absent from region_extents) survived the rebuild.

src/app.js, src/app.html, tests/js/smoke.spec.js and
tests/python/test_atlaslib.py all merged automatically - this branch
and the fissures/compound-labels/nifti-download work never touch the
same code.

tests/python: 29 passed / 1 failed (missing skimage on this machine's
global Python), the documented clean baseline plus this branch's own
new test. tests/js could not run locally (no node here) - CI is
authoritative, including for the Tokyo-timezone assertion that has
never executed anywhere yet.

+++ tools/volume.py
@@ -324,7 +324,11 @@ def write_nifti(path, labels, grid, descrip='', unnamed=65000):
     written beside it. Nothing here needs nibabel: the header is 348 bytes of struct."""
     import gzip
     import struct
-    zyx = np.where(labels == UNASSIGNED, unnamed, labels).astype(np.uint16)   # (AP, DV, ML)
+    # cast before the where, not after: `unnamed` is 65000 and `labels` is int16, and
+    # numpy 2 refuses a Python int the array's dtype cannot hold rather than widening
+    # the result. The mask is read off the signed array, so UNASSIGNED still finds it.
+    zyx = np.where(labels == UNASSIGNED, np.uint16(unnamed),
+                   labels.astype(np.uint16))                              # (AP, DV, ML)
     arr = np.ascontiguousarray(zyx.transpose(2, 0, 1)[:, ::-1, :])        # (ML, AP asc, DV)
     nx, ny, nz = arr.shape
     res = float(grid.res)
